Fixes broken app/link under non-Windows

+++ build/controllers/AppController.php
@@ -29,7 +29,7 @@ class AppController extends Controller
      */
     protected function unlink($file)
     {
-        if (is_dir($file)) {
+        if (is_dir($file) && strncasecmp(PHP_OS, 'WIN', 3) === 0) {
             rmdir($file);
         } else {
             unlink($file);
